Week 6/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2

# Data processing libraries
pandas==2.1.3
openpyxl==3.1.2
tiktoken==0.6.0

# Web scraping and HTML processing
selenium==4.15.2
beautifulsoup4==4.12.2
html2text==2020.1.16
readability-lxml==0.8.1

# AI and LLM integration
openai==1.15.0
langchain==0.1.9
langchain-core==0.1.27
langchain-openai==0.0.9
langchain-community==0.0.29

# Database
pymongo==4.6.1
dnspython==2.5.0

# UI and visualization
gradio==4.26.0
